research_aggregator/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
torch==2.2.1
nltk==3.8.1
pandas==2.1.4
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
newspaper3k==0.2.8
scikit-learn==1.3.2
spacy==3.7.2
tqdm==4.66.1
numpy==1.26.0
gensim==4.3.2
sentence-transformers==2.0.0
redis==5.0.1
pymongo==4.6.1
aiohttp==3.9.1
asyncio==3.4.3
lxml[html_clean]
huggingface_hub>=0.19.3
PyPDF2==3.0.1
streamlit==1.36.0
mistral_inference==1.6.0
mistral_common==1.6.2
